Create tiingo_refresh.gs
This is the same code as you find in tiingo.gs. However, I've added a janky workaround to permit manual refreshing in Google Sheets through the addition of an unused parameter. I can see how this isn't a great solution and isn't really clean code for those that would prefer a clean code base. Hopefully, Google will permit an easier time of creating refresh intervals, but I doubt this is functionality that it desires to permit in a "free" product. If that functionality is created, this branch can be retired. Until then, this will continue in parallel to the main, "clean", branch with the janky workaround included.